% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fsef_e10.R
\name{calculate_fsef_e10}
\alias{calculate_fsef_e10}
\title{Calculate the fuel specific emission factors for all pollutants using
e10 petrol parameters.}
\usage{
calculate_fsef_e10(df)
}
\arguments{
\item{df}{The data frame containing remote sensing pollutant emission ratios}
}
\value{
The fuel specific emission factors for all pollutants using e10 petrol
parameters.
}
\description{
The calculate_fsef_std calculates the fuel specific emission
factor for the pollutants CO2, CO, HC, NO, NO2, NOx, and NH3 using the
following parameters: molar mass of fuel MF = 14.544, response factor RF =
2.0, scaling factor SF = 3.0.
}
