% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fsef_std.R
\name{calculate_fsef_std}
\alias{calculate_fsef_std}
\title{Calculate the fuel specific emission factors for all pollutants using
standard parameters.}
\usage{
calculate_fsef_std(df)
}
\arguments{
\item{df}{The data frame containing remote sensing pollutant emission ratios}
}
\value{
The fuel specific emission factors for all pollutants using standard
parameters.
}
\description{
The calculate_fsef_std calculates the fuel specific emission
factor for the pollutants CO2, CO, HC, NO, NO2, NOx, and NH3 using the
following parameters: molar mass of fuel MF = 14.000, response factor RF =
2.0, scaling factor SF = 3.0.
}
