% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e5_fsef.R
\name{e5_fsef_co}
\alias{e5_fsef_co}
\alias{e5_fsef_hc}
\alias{e5_fsef_no_as_no2}
\alias{e5_fsef_no}
\alias{e5_fsef_no2}
\alias{e5_fsef_nh3}
\title{Calculate the fuel specific emission factor for a pollutant using a fuel
mass value of 14.384}
\usage{
e5_fsef_co(ratio_co_co2, ratio_hc_co2)

e5_fsef_hc(ratio_hc_co2, ratio_co_co2)

e5_fsef_no_as_no2(ratio_no_co2, ratio_co_co2, ratio_hc_co2)

e5_fsef_no(ratio_no_co2, ratio_co_co2, ratio_hc_co2)

e5_fsef_no2(ratio_no2_co2, ratio_co_co2, ratio_hc_co2)

e5_fsef_nh3(ratio_nh3_co2, ratio_co_co2, ratio_hc_co2)
}
\arguments{
\item{ratio_co_co2}{The ratio of carbon monoxide to carbon dioxide}

\item{ratio_hc_co2}{The ratio of un-burnt hydrocarbons to carbon dioxide}

\item{ratio_no_co2}{The ratio of nitrogen monoxide to carbon dioxide}

\item{ratio_no2_co2}{The ratio of nitrogen dioxide to carbon dioxide}

\item{ratio_nh3_co2}{The ratio of ammonia to carbon dioxide}
}
\value{
The fuel specific emission factor for the pollutant calculated using
a fuel mass value of 14.384
}
\description{
The e5_fsef_x functions calculate the fuel specific emission factor for
the pollutant x using the measured pollutant ratios to carbon dioxide, assuming
a fuel mass of 14.384 (E5 petrol).
}
