% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsensing.R
\name{genLUT}
\alias{genLUT}
\title{genLUT}
\usage{
genLUT(...)
}
\arguments{
\item{...}{At least 2 parameters or a list composed by 2 parameters are needed. It could be an array of numbers or characters.}
}
\value{
a matrix which is indeed the Lookup Table
}
\description{
\code{genLUT} generates a Lookup table with given parameters. This is very useful when you are generating a data entry for
RTM inversion
}
\examples{
genLUT(Cab = 10:30, LAI = seq(0, 0.05, 0.01), Car = runif(10, 2.0, 3.5))
}
