% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsensing.R
\name{long2UTM}
\alias{long2UTM}
\title{long2UTM}
\usage{
long2UTM(long)
}
\arguments{
\item{long}{Numeric. Longitude coordinate}
}
\description{
\code{long2UTM} converts the given longitude into the corresponing UTM zone.
}
\examples{
long2UTM(-4)

}
