% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsensing.R
\name{polygonize}
\alias{polygonize}
\title{polygonize}
\usage{
polygonize(raster)
}
\arguments{
\item{raster}{Raster. The binary raster object to be converted.}
}
\value{
a spatial object.
}
\description{
\code{polygonize} is a faster alternative to \link{rasterToPolygons}.
}
\details{
This is an alternative function to \link{rasterToPolygons}, which pretends to be
the fastest way to perform this task. Original idea from \code{gdal_polygonizeR()}
It also offers a directly way to create a mask from a binary raster.

It has a dependency on GDAL Tools to be installed
}
\examples{
\dontrun{
shape <- polygonize(raster)
}
}
