% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsensing.R
\name{rasterToPolygons0}
\alias{rasterToPolygons0}
\title{rasterToPolygons0}
\usage{
rasterToPolygons0(raster, value2keep = "")
}
\arguments{
\item{raster}{the raster object to be converted.}

\item{value2keep}{the path where it should be written.}
}
\value{
a spatial object
}
\description{
\code{rasterToPolygons0} is a faster alternative to \link{rasterToPolygons}.
}
\details{
This is an alternative function to \link{rasterToPolygons}, which pretends to be
the fastest way to perform this task. Original idea from \code{gdal_polygonizeR()}
It also offers a directly way to create a mask.

It has a dependency on GDAL Tools to be installed
}
\examples{
shape <- rasterToPolygons0(rasterObject)
shape <- rasterToPolygons0(rasterObject, 255)

}
