% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsensing.R
\name{read.shape}
\alias{read.shape}
\title{read.shape}
\usage{
read.shape(shape.path)
}
\arguments{
\item{shape.path}{the path of the shapefile.}
}
\description{
\code{read.shape} is an easier alternative to \link{readOGR}.
}
\details{
This is an alternative function to \link{readOGR}, which pretends to be the easiest way
to load shapefiles.
}
\examples{
read.shape('./myshapefile.shp')

}
