% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsensing.R
\name{shape2grid}
\alias{shape2grid}
\title{shape2grid}
\usage{
shape2grid(shape, gsize = 100)
}
\arguments{
\item{shape}{Shapefile. Spatial object to be converted}

\item{gsize}{Optional. Numeric. Size of each cell in meters. 100 m by default.}
}
\value{
a spatial object.
}
\description{
\code{shape2grid} converts a shape into a grid. It is very useful for
clustering purposes.
}
\examples{
\dontrun{
shape <- shape2grid(shape, 100)
}
}
