% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsensing.R
\name{shapeInterpolation}
\alias{shapeInterpolation}
\title{shapeInterpolation}
\usage{
shapeInterpolation(shape, field, cellsize, shape.mask = NULL,
  showPercentage = FALSE)
}
\arguments{
\item{shape}{Shapefile. Spatial object to be interpolated}

\item{field}{Character. Field to be considered}

\item{cellsize}{Numeric. Raster output resolution in meters}

\item{shape.mask}{Optional. Shapefile object as a mask}

\item{showPercentage}{Optional. Logical. By default no percentage is shown}
}
\value{
a raster.
}
\description{
\code{shapeInterpolation} interpolates a shapefile field to a raster with \link{krige}
}
\examples{
\dontrun{
raster.interpolated <- shapeInterpolation(shape2, 'NDVI', 0.2, showPercentage = T)
raster.interpolated <- shapeInterpolation(shape2, 'TC', 5, shape.mask = shapemask)
}
}
