% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsensing.R
\name{stats2perform}
\alias{stats2perform}
\title{stats2perform}
\usage{
stats2perform(x, na.rm = FALSE)
}
\arguments{
\item{x}{set of data to summarised}

\item{na.rm}{Optional. Logical. If na.rm=TRUE, NA values are removed}
}
\value{
statistics. Mean, min, max, sd and number of pixels.
}
\description{
\code{stats2perform} defines the default stats for \link{extractRasterLayer} and \link{extractRasterStack}
}
