% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsensing.R
\name{write.shape}
\alias{write.shape}
\title{write.shape}
\usage{
write.shape(shape, shape.path)
}
\arguments{
\item{shape}{the spatial object to be saved.}

\item{shape.path}{the path where it should be written.}
}
\description{
\code{write.shape} is an easier alternative to \link{writeOGR}.
}
\details{
This is an alternative function to \link{writeOGR}, which pretends to be the easiest way
to write shapefiles.
}
\examples{
write.shape(shapeObject, './myshapefile.shp')

}
