% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsensing.R
\name{gdal_fillnodata}
\alias{gdal_fillnodata}
\title{gdal_fillnodata}
\usage{
gdal_fillnodata(raster, mask)
}
\arguments{
\item{raster}{Raster. The raster object to be processed, just one band.}

\item{mask}{Raster. The raster mask with zeroes and ones. It should be a binary mask.}
}
\description{
\code{gdal_fillnodata} is a tiny \code{gdal_fillnodata.py} wrapper from the GDAL library tools.
}
\details{
The aim is just to convert the given raster and mask into a filled raster.

It has a dependency on GDAL Tools to be installed
}
\examples{
\dontrun{
gdal_fillnodata(raster, binaryMask)
}
}
